# swagger_server/models/new_topology_nodes.py
""" class new topology nodes """
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.location import Location  # noqa: F401,E501
from swagger_server.models.port import Port  # noqa: F401,E501
import re  # noqa: F401,E501
from swagger_server import util


class NewTopologyNodes(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, location: Location=None, ports: List[Port]=None):  # noqa: E501
        """NewTopologyNodes - a model defined in Swagger

        :param id: The id of this NewTopologyNodes.  # noqa: E501
        :type id: str
        :param name: The name of this NewTopologyNodes.  # noqa: E501
        :type name: str
        :param location: The location of this NewTopologyNodes.  # noqa: E501
        :type location: Location
        :param ports: The ports of this NewTopologyNodes.  # noqa: E501
        :type ports: List[Port]
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'location': Location,
            'ports': List[Port]
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'location': 'location',
            'ports': 'ports'
        }
        self._id = id
        self._name = name
        self._location = location
        self._ports = ports

    @classmethod
    def from_dict(cls, dikt) -> 'NewTopologyNodes':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The NewTopology_nodes of this NewTopologyNodes.  # noqa: E501
        :rtype: NewTopologyNodes
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this NewTopologyNodes.


        :return: The id of this NewTopologyNodes.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this NewTopologyNodes.


        :param id: The id of this NewTopologyNodes.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this NewTopologyNodes.


        :return: The name of this NewTopologyNodes.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this NewTopologyNodes.


        :param name: The name of this NewTopologyNodes.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def location(self) -> Location:
        """Gets the location of this NewTopologyNodes.


        :return: The location of this NewTopologyNodes.
        :rtype: Location
        """
        return self._location

    @location.setter
    def location(self, location: Location):
        """Sets the location of this NewTopologyNodes.


        :param location: The location of this NewTopologyNodes.
        :type location: Location
        """
        if location is None:
            raise ValueError("Invalid value for `location`, must not be `None`")  # noqa: E501

        self._location = location

    @property
    def ports(self) -> List[Port]:
        """Gets the ports of this NewTopologyNodes.


        :return: The ports of this NewTopologyNodes.
        :rtype: List[Port]
        """
        return self._ports

    @ports.setter
    def ports(self, ports: List[Port]):
        """Sets the ports of this NewTopologyNodes.


        :param ports: The ports of this NewTopologyNodes.
        :type ports: List[Port]
        """
        if ports is None:
            raise ValueError("Invalid value for `ports`, must not be `None`")  # noqa: E501

        self._ports = ports